Address review: remove Bf16 codec, extract sq8_simd module
Two changes bundled:

1. Drop the `RerankCodec::Bf16` variant. It missed the 0.85
   reservoir-recall floor (recall ~ 0.775) — the query-side bf16
   rounding noise stacks on top of the doc-side rounding, and the
   2x space win is dominated by Sq8's 4x win at higher recall.
   - On-disk `codec_id = 1` is retired (segments that used it can
     no longer be opened — Bf16 never shipped to main).
   - Builder match arm, all `distance::*bf16*` kernels + their
     SIMD parity tests, `simd_dispatch::has_bf16_dot`, and the
     parametric reservoir-recall sweep entry are removed.
   - Plan 012 carries a removal note at the top.

2. Address Pranav's review on #24:
   - Extract Sq8 SIMD primitives (`update_min_max`,
     `sq8_encode_row` + their wide/AVX2/AVX-512 tier impls, the
     `Sq8EncodeConsts` precompute) and the SIMD parity / microbench
     tests into a new `sq8_simd` module. `builder.rs` keeps the
     build orchestration (`encode_sq8_cluster_simd`,
     `derive_sq8_quantizer_from_min_max`) and calls into the new
     module via three `pub(super)` entry points.
   - `Sq8EncodeConsts::from_scale_offset` builds `inv_scale` / `c2`
     via iterator chains instead of `vec![0.0; dim]` + index
     assignment.
   - Doc-comment in the AVX2 encode kernel explains the
     `_mm256_extracti128_si256::<1>(...)` turbofish (const-generic
     imm8 in the Rust binding) and the packing rationale (SIMD
     equivalent of the wide path's per-lane scalar narrow, with
     one 8-byte store instead of eight 1-byte stores).
   - Adds `sq8_encode_consts_match_algebraic_identity` unit test
     pinning the precompute against `inv = 1/scale`,
     `c2 = (-offset) * inv + 0.5`.

All 148 `superfile::vector::*` lib tests pass; the reservoir-recall
integration test (which originally flagged the Bf16 regression)
passes on the surviving Fp32 + Sq8 codecs.
