+++ src/superfile/vector/builder.rs
@@ -17,7 +17,7 @@ use crate::superfile::vector::quant::BitQuantizer;
 use crate::superfile::vector::rerank_codec::RerankCodec;
 use crate::superfile::vector::reservoir::{Reservoir, default_kmeans_sample_size};
 use crate::superfile::vector::rotation::RandomRotation;
-use crate::superfile::vector::simd_dispatch::{avx2_enabled, avx512_enabled};
+use crate::superfile::vector::sq8_simd::{Sq8EncodeConsts, sq8_encode_row, update_min_max};
 use crate::superfile::vector::spill::{
     ChunkedVectorSource, InMemoryVectorSource, MmapVectorSource, SpillWriter,
 };
@@ -839,7 +839,7 @@ fn build_subsection_streaming(
     let codes_off = cluster_idx_off + cluster_idx_size;
     // `codec_meta_off` is the on-disk slot in the sub-header for the
     // codec metadata region's start. Zero means "no codec metadata
-    // in this subsection" — Fp32 / Bf16 / RabitqOnly write 0 here
+    // in this subsection" — Fp32 / RabitqOnly write 0 here
     // and stay byte-identical to legacy fp32 segments that left the
     // (formerly reserved) slot zero.
     let codec_meta_off_value: u32 = if codec_meta_size == 0 {
@@ -981,7 +981,6 @@ fn build_subsection_streaming(
     let mut id_block: Vec<u8> = Vec::new();
     let mut code_block: Vec<u8> = Vec::new();
     let mut full_block: Vec<u8> = Vec::new();
-    let mut row_fp32_scratch: Vec<f32> = Vec::new();
 
     for (c, &(cluster_off_u32, cluster_count_u32)) in cluster_index.iter().enumerate() {
         if cluster_count_u32 == 0 {
@@ -1023,34 +1022,13 @@ fn build_subsection_streaming(
 
         // full[] region — codec-dependent. RabitqOnly has no full[]
         // payload on disk and pass 2 didn't spill one. Fp32 is a
-        // direct block copy. Bf16/Sq8 transcode out of the fp32
-        // block buffer.
+        // direct block copy. Sq8 transcodes out of the fp32 block
+        // buffer.
         match codec {
             RerankCodec::RabitqOnly => {}
             RerankCodec::Fp32 => {
                 let dst_base = full_off + cluster_off * dim * 4;
                 bytes[dst_base..dst_base + cluster_count * dim * 4].copy_from_slice(&full_block);
-            }
-            RerankCodec::Bf16 => {
-                // bf16 destination is 2 bytes/lane; read the fp32
-                // block as &[f32] (Vec<u8> is f32-aligned only when
-                // we go through bytemuck on a fresh aligned slab —
-                // here we re-stage one row at a time to keep lane-
-                // wise rounding identical to the previous code path).
-                if row_fp32_scratch.len() < dim {
-                    row_fp32_scratch.resize(dim, 0.0);
-                }
-                for i in 0..cluster_count {
-                    let pos = cluster_off + i;
-                    let src = &full_block[i * dim * 4..(i + 1) * dim * 4];
-                    let row = &mut row_fp32_scratch[..dim];
-                    bytemuck::cast_slice_mut::<f32, u8>(row).copy_from_slice(src);
-                    let off = full_off + pos * dim * 2;
-                    for (d, &x) in row.iter().enumerate() {
-                        let bf = crate::superfile::vector::distance::fp32_to_bf16(x);
-                        bytes[off + d * 2..off + d * 2 + 2].copy_from_slice(&bf.to_le_bytes());
-                    }
-                }
             }
             RerankCodec::Sq8 => {
                 // The cluster's fp32 rows are already in `full_block`
@@ -1183,427 +1161,6 @@ fn derive_sq8_quantizer_from_min_max(min: &[f32], max: &[f32]) -> (Vec<f32>, Vec
     (scale, offset)
 }
 
-/// Per-row, per-cluster `(min, max)` update for the Sq8 quantizer.
-/// Called from pass 2 for every fp32 row routed to its bucket: takes
-/// the row and the destination cluster's `min[dim]` / `max[dim]`
-/// slices and updates them in-place with the per-dim element-wise
-/// `min`/`max`.
-///
-/// Folds what used to be pass 3's per-cluster min/max scan into the
-/// pass that already touches every row anyway, eliminating one
-/// full re-read of the cluster's fp32 bytes after pass 2 spills them.
-/// Three-tier dispatch: AVX-512 (16-lane `vminps` + `vmaxps`) →
-/// AVX2 (8-lane `vminps` + `vmaxps`) → portable `wide::f32x8`.
-#[inline]
-fn update_min_max(row: &[f32], min_slice: &mut [f32], max_slice: &mut [f32]) {
-    debug_assert_eq!(row.len(), min_slice.len());
-    debug_assert_eq!(row.len(), max_slice.len());
-
-    #[cfg(target_arch = "x86_64")]
-    {
-        if avx512_enabled() {
-            // SAFETY: gated on avx512_enabled() which requires `avx512f`.
-            unsafe { update_min_max_avx512(row, min_slice, max_slice) };
-            return;
-        }
-        if avx2_enabled() {
-            // SAFETY: gated on avx2_enabled() which requires `avx2`.
-            unsafe { update_min_max_avx2(row, min_slice, max_slice) };
-            return;
-        }
-    }
-    update_min_max_wide(row, min_slice, max_slice);
-}
-
-/// Portable `wide::f32x8` (256-bit) per-dim min/max update. Universal
-/// fallback for aarch64 / SSE-only x86_64 / `INFINO_DISABLE_AVX2=1`.
-#[inline]
-fn update_min_max_wide(row: &[f32], min_slice: &mut [f32], max_slice: &mut [f32]) {
-    use wide::f32x8;
-    let dim = row.len();
-    let full = dim - dim % 8;
-    let mut i = 0;
-    while i < full {
-        let r: [f32; 8] = row[i..i + 8].try_into().expect("len 8");
-        let mn: [f32; 8] = min_slice[i..i + 8].try_into().expect("len 8");
-        let mx: [f32; 8] = max_slice[i..i + 8].try_into().expect("len 8");
-        let r_v = f32x8::from(r);
-        let new_min = r_v.fast_min(f32x8::from(mn)).to_array();
-        let new_max = r_v.fast_max(f32x8::from(mx)).to_array();
-        min_slice[i..i + 8].copy_from_slice(&new_min);
-        max_slice[i..i + 8].copy_from_slice(&new_max);
-        i += 8;
-    }
-    while i < dim {
-        let x = row[i];
-        if x < min_slice[i] {
-            min_slice[i] = x;
-        }
-        if x > max_slice[i] {
-            max_slice[i] = x;
-        }
-        i += 1;
-    }
-}
-
-/// AVX2 per-dim min/max update. 8 lanes per iteration with
-/// `_mm256_min_ps` + `_mm256_max_ps` (single-instruction parallel
-/// reduce). Lifts every AVX2 host that lacks AVX-512.
-///
-/// # Safety
-///
-/// Callers must ensure the target supports `avx2`.
-#[cfg(target_arch = "x86_64")]
-#[target_feature(enable = "avx2")]
-unsafe fn update_min_max_avx2(row: &[f32], min_slice: &mut [f32], max_slice: &mut [f32]) {
-    use std::arch::x86_64::*;
-    let dim = row.len();
-    let full = dim - dim % 8;
-    let mut i = 0;
-    // SAFETY: each iteration reads 24 bytes (one f32 lane of row,
-    // min, max each at offset `i`) and writes 16 bytes (min, max
-    // at the same offset). `i + 8 <= dim` keeps every load/store
-    // in bounds; loads/stores are all unaligned.
-    unsafe {
-        while i < full {
-            let r = _mm256_loadu_ps(row.as_ptr().add(i));
-            let mn = _mm256_loadu_ps(min_slice.as_ptr().add(i));
-            let mx = _mm256_loadu_ps(max_slice.as_ptr().add(i));
-            let new_mn = _mm256_min_ps(r, mn);
-            let new_mx = _mm256_max_ps(r, mx);
-            _mm256_storeu_ps(min_slice.as_mut_ptr().add(i), new_mn);
-            _mm256_storeu_ps(max_slice.as_mut_ptr().add(i), new_mx);
-            i += 8;
-        }
-    }
-    while i < dim {
-        let x = row[i];
-        if x < min_slice[i] {
-            min_slice[i] = x;
-        }
-        if x > max_slice[i] {
-            max_slice[i] = x;
-        }
-        i += 1;
-    }
-}
-
-/// AVX-512 per-dim min/max update. 16 lanes per iteration with
-/// `_mm512_min_ps` + `_mm512_max_ps`. Strictly faster than the AVX2
-/// path on Sapphire Rapids / Granite Rapids / Zen 4+ because every
-/// reduction step processes twice the lanes per cycle.
-///
-/// # Safety
-///
-/// Callers must ensure the target supports `avx512f`.
-#[cfg(target_arch = "x86_64")]
-#[target_feature(enable = "avx512f")]
-unsafe fn update_min_max_avx512(row: &[f32], min_slice: &mut [f32], max_slice: &mut [f32]) {
-    use std::arch::x86_64::*;
-    let dim = row.len();
-    let full16 = dim - dim % 16;
-    let mut i = 0;
-    // SAFETY: per-iteration windows of 16 f32s in each of row /
-    // min / max; bounded by `i + 16 <= dim`. Unaligned loads /
-    // stores.
-    unsafe {
-        while i < full16 {
-            let r = _mm512_loadu_ps(row.as_ptr().add(i));
-            let mn = _mm512_loadu_ps(min_slice.as_ptr().add(i));
-            let mx = _mm512_loadu_ps(max_slice.as_ptr().add(i));
-            let new_mn = _mm512_min_ps(r, mn);
-            let new_mx = _mm512_max_ps(r, mx);
-            _mm512_storeu_ps(min_slice.as_mut_ptr().add(i), new_mn);
-            _mm512_storeu_ps(max_slice.as_mut_ptr().add(i), new_mx);
-            i += 16;
-        }
-        // 8-lane half-tail for `dim % 16 == 8` (rare but
-        // possible — keeps the kernel correct without falling to
-        // scalar on the same iteration count).
-        if i + 8 <= dim {
-            let r = _mm256_loadu_ps(row.as_ptr().add(i));
-            let mn = _mm256_loadu_ps(min_slice.as_ptr().add(i));
-            let mx = _mm256_loadu_ps(max_slice.as_ptr().add(i));
-            let new_mn = _mm256_min_ps(r, mn);
-            let new_mx = _mm256_max_ps(r, mx);
-            _mm256_storeu_ps(min_slice.as_mut_ptr().add(i), new_mn);
-            _mm256_storeu_ps(max_slice.as_mut_ptr().add(i), new_mx);
-            i += 8;
-        }
-    }
-    while i < dim {
-        let x = row[i];
-        if x < min_slice[i] {
-            min_slice[i] = x;
-        }
-        if x > max_slice[i] {
-            max_slice[i] = x;
-        }
-        i += 1;
-    }
-}
-
-/// Per-cluster Sq8 encode constants. Each cluster contributes one
-/// `(inv_scale[dim], c2[dim])` pair where
-///     `c2[d] = (-offset[d]) * inv_scale[d] + 0.5`
-/// so the per-row encode collapses to a single FMA:
-///     `q = x * inv_scale + c2 = (x - offset) * inv_scale + 0.5`
-/// followed by `clamp(0, 255)` and a truncating cast to `u8`. The
-/// `+ 0.5` is what turns the truncating cast into round-half-up,
-/// which equals round-half-away on the non-negative domain that
-/// the Sq8 encoder operates in (the float result is in `[0, 255]`
-/// for in-cluster values; clamp catches the fp-noise overshoot at
-/// both ends).
-///
-/// Precomputing per cluster (rather than per row) hoists `1/scale`
-/// and the constant fold out of the dim-axis inner loop, so the
-/// SIMD encoder loops over a single FMA + cast + narrow per lane.
-struct Sq8EncodeConsts {
-    inv_scale: Vec<f32>,
-    c2: Vec<f32>,
-}
-
-impl Sq8EncodeConsts {
-    fn from_scale_offset(scale: &[f32], offset: &[f32]) -> Self {
-        debug_assert_eq!(scale.len(), offset.len());
-        let dim = scale.len();
-        let mut inv_scale = vec![0.0f32; dim];
-        let mut c2 = vec![0.0f32; dim];
-        for d in 0..dim {
-            inv_scale[d] = 1.0 / scale[d];
-            c2[d] = (-offset[d]).mul_add(inv_scale[d], 0.5);
-        }
-        Self { inv_scale, c2 }
-    }
-}
-
-/// SIMD f32 → u8 Sq8 encode. Writes one row's `dim` codes to
-/// `dst[0..dim]` from `row[0..dim]` and the cluster's precomputed
-/// `inv_scale` / `c2` arrays.
-///
-/// All three tiers use the same FMA + clamp + truncating-int-cast
-/// sequence so the byte output is bit-identical across paths:
-///
-///   q = FMA(x, inv_scale, c2)
-///   q_clamped = max(0, min(255, q))
-///   dst[d] = (q_clamped) as u8     // truncating cast
-///
-/// The scalar fallback in the tail loop uses `f32::mul_add` to keep
-/// FMA semantics consistent with the SIMD lanes — `(x * inv) + c2`
-/// without FMA would double-round and produce a ≤1 ulp drift from
-/// the SIMD result, which is enough to flip a quantization boundary
-/// and change the on-disk byte for that lane.
-#[inline]
-fn sq8_encode_row(row: &[f32], inv_scale: &[f32], c2: &[f32], dst: &mut [u8]) {
-    debug_assert_eq!(row.len(), inv_scale.len());
-    debug_assert_eq!(row.len(), c2.len());
-    debug_assert_eq!(row.len(), dst.len());
-
-    #[cfg(target_arch = "x86_64")]
-    {
-        if avx512_enabled() {
-            // SAFETY: gated on `avx512f` feature detection.
-            unsafe { sq8_encode_row_avx512(row, inv_scale, c2, dst) };
-            return;
-        }
-        if avx2_enabled() {
-            // SAFETY: gated on `avx2` feature detection.
-            unsafe { sq8_encode_row_avx2(row, inv_scale, c2, dst) };
-            return;
-        }
-    }
-    sq8_encode_row_wide(row, inv_scale, c2, dst);
-}
-
-/// Portable Sq8 encode. `wide::f32x8` handles the FMA + clamp in
-/// SIMD; the narrow to u8 falls back to a per-lane scalar cast
-/// because `wide` doesn't expose a saturating u8 narrow. Still
-/// vectorizes the dominant fp work.
-//
-// `q.max(0).min(255)` instead of `q.clamp(0, 255)`: matches the
-// SIMD `_mm{256,512}_max_ps(_mm{256,512}_min_ps(q, 255), 0)` NaN
-// handling exactly so the wide and AVX{2,512} tiers produce
-// bit-identical bytes. clamp() returns NaN for NaN input where
-// the SIMD intrinsics return one of the limits; FMA on finite
-// inputs is always finite so the difference is unreachable in
-// the encode pipeline, but the parity tests assert byte-equality
-// across all tiers and the two-step formulation keeps that
-// pre-condition independent of the input domain.
-#[allow(clippy::manual_clamp)]
-fn sq8_encode_row_wide(row: &[f32], inv_scale: &[f32], c2: &[f32], dst: &mut [u8]) {
-    use wide::f32x8;
-    let dim = row.len();
-    let zero = f32x8::splat(0.0);
-    let max255 = f32x8::splat(255.0);
-    let mut i = 0;
-    while i + 8 <= dim {
-        let r: [f32; 8] = row[i..i + 8].try_into().expect("len 8");
-        let inv: [f32; 8] = inv_scale[i..i + 8].try_into().expect("len 8");
-        let c: [f32; 8] = c2[i..i + 8].try_into().expect("len 8");
-        let q = f32x8::from(r).mul_add(f32x8::from(inv), f32x8::from(c));
-        let q_clamped = q.fast_max(zero).fast_min(max255).to_array();
-        for k in 0..8 {
-            dst[i + k] = q_clamped[k] as u8;
-        }
-        i += 8;
-    }
-    while i < dim {
-        let q = row[i].mul_add(inv_scale[i], c2[i]);
-        let q_clamped = q.max(0.0).min(255.0);
-        dst[i] = q_clamped as u8;
-        i += 1;
-    }
-}
-
-/// AVX2 Sq8 encode. 8 lanes per iteration: `VFMADD213PS` →
-/// `VMINPS`/`VMAXPS` clamp → `VCVTTPS2DQ` truncate-to-i32 →
-/// `VPACKUSDW` (i32 → u16 saturating) → `VPACKUSWB`
-/// (u16 → u8 saturating) → 8-byte store. The saturating narrow
-/// is a no-op because we already clamped to `[0, 255]` in float
-/// space; it's the cheapest path that produces packed u8.
-///
-/// # Safety
-///
-/// Callers must ensure the target supports `avx2` and `fma`.
-#[cfg(target_arch = "x86_64")]
-#[target_feature(enable = "avx2,fma")]
-#[allow(clippy::manual_clamp)] // see sq8_encode_row_wide note
-unsafe fn sq8_encode_row_avx2(row: &[f32], inv_scale: &[f32], c2: &[f32], dst: &mut [u8]) {
-    use std::arch::x86_64::*;
-    let dim = row.len();
-    let zero = _mm256_setzero_ps();
-    let max255 = _mm256_set1_ps(255.0);
-    let mut i = 0;
-    // SAFETY: each iteration reads 8 f32 lanes from `row`,
-    // `inv_scale`, `c2` at offset `i` and writes 8 bytes to
-    // `dst[i..i+8]`. `i + 8 <= dim` bounds every load and store.
-    unsafe {
-        while i + 8 <= dim {
-            let r = _mm256_loadu_ps(row.as_ptr().add(i));
-            let inv = _mm256_loadu_ps(inv_scale.as_ptr().add(i));
-            let c = _mm256_loadu_ps(c2.as_ptr().add(i));
-            // q = x * inv + c  (single rounding)
-            let q = _mm256_fmadd_ps(r, inv, c);
-            // clamp to [0, 255]
-            let q_clamped = _mm256_max_ps(_mm256_min_ps(q, max255), zero);
-            // truncate-to-i32; lanes are in [0, 255] so the i32
-            // values are non-negative and PACKUSDW interprets them
-            // as unsigned correctly.
-            let q_i32 = _mm256_cvttps_epi32(q_clamped);
-            // Narrow 8 × i32 → 8 × u8 via two saturating packs.
-            let lo = _mm256_castsi256_si128(q_i32);
-            let hi = _mm256_extracti128_si256::<1>(q_i32);
-            let packed_u16 = _mm_packus_epi32(lo, hi); // 8 × u16
-            let packed_u8 = _mm_packus_epi16(packed_u16, packed_u16); // low 8 bytes valid
-            // Store the low 64 bits (8 bytes). Unaligned.
-            let dst_ptr = dst.as_mut_ptr().add(i) as *mut i64;
-            std::ptr::write_unaligned(dst_ptr, _mm_cvtsi128_si64(packed_u8));
-            i += 8;
-        }
-    }
-    while i < dim {
-        let q = row[i].mul_add(inv_scale[i], c2[i]);
-        let q_clamped = q.max(0.0).min(255.0);
-        dst[i] = q_clamped as u8;
-        i += 1;
-    }
-}
-
-/// AVX-512 Sq8 encode. 16 lanes per iteration: `VFMADD213PS` →
-/// `VMINPS`/`VMAXPS` clamp → `VCVTTPS2DQ` truncate-to-i32 →
-/// `VPMOVUSDB` (single-instruction unsigned-saturating narrow
-/// i32 → u8) → 16-byte store. The narrow is one instruction
-/// instead of the two pack steps the AVX2 path needs, which is
-/// where the AVX-512 path picks up its extra factor over AVX2
-/// on top of the 2× lane count.
-///
-/// # Safety
-///
-/// Callers must ensure the target supports `avx512f`.
-#[cfg(target_arch = "x86_64")]
-#[target_feature(enable = "avx512f")]
-#[allow(clippy::manual_clamp)] // see sq8_encode_row_wide note
-unsafe fn sq8_encode_row_avx512(row: &[f32], inv_scale: &[f32], c2: &[f32], dst: &mut [u8]) {
-    use std::arch::x86_64::*;
-    let dim = row.len();
-    let zero = _mm512_setzero_ps();
-    let max255 = _mm512_set1_ps(255.0);
-    let mut i = 0;
-    // SAFETY: each iteration reads 16 f32 lanes from `row`,
-    // `inv_scale`, `c2` at offset `i` and writes 16 bytes to
-    // `dst[i..i+16]`. `i + 16 <= dim` bounds every load and store.
-    unsafe {
-        while i + 16 <= dim {
-            let r = _mm512_loadu_ps(row.as_ptr().add(i));
-            let inv = _mm512_loadu_ps(inv_scale.as_ptr().add(i));
-            let c = _mm512_loadu_ps(c2.as_ptr().add(i));
-            let q = _mm512_fmadd_ps(r, inv, c);
-            let q_clamped = _mm512_max_ps(_mm512_min_ps(q, max255), zero);
-            let q_i32 = _mm512_cvttps_epi32(q_clamped);
-            // Single-instruction unsigned-saturating narrow i32 → u8.
-            // Result lanes are guaranteed in [0, 255] so the
-            // saturation is a no-op; we use the unsigned variant
-            // because it interprets the source as u32 (the i32
-            // values are non-negative so the bit pattern matches).
-            let packed_u8 = _mm512_cvtusepi32_epi8(q_i32); // __m128i (16 bytes)
-            _mm_storeu_si128(dst.as_mut_ptr().add(i) as *mut __m128i, packed_u8);
-            i += 16;
-        }
-    }
-    // 8-lane half-tail for `dim % 16 == 8`.
-    #[cfg(target_arch = "x86_64")]
-    if i + 8 <= dim {
-        // SAFETY: target_feature `avx512f` implies `avx2` + `fma`
-        // are available; we just call the AVX2 path on the
-        // remaining 8 lanes (no need to re-emit the same code).
-        unsafe { sq8_encode_row_avx2_unsafe_tail8(row, inv_scale, c2, dst, &mut i) };
-    }
-    while i < dim {
-        let q = row[i].mul_add(inv_scale[i], c2[i]);
-        let q_clamped = q.max(0.0).min(255.0);
-        dst[i] = q_clamped as u8;
-        i += 1;
-    }
-}
-
-/// 8-lane Sq8 encode tail used by the AVX-512 kernel when
-/// `dim % 16 == 8`. Same FMA + clamp + pack sequence as
-/// `sq8_encode_row_avx2` but inlined so the AVX-512 entry point
-/// keeps target_feature("avx512f") and we don't pay a tail call.
-///
-/// # Safety
-///
-/// Callers must ensure `avx2` + `fma` are available, and that
-/// `*i + 8 <= row.len()`.
-#[cfg(target_arch = "x86_64")]
-#[target_feature(enable = "avx2,fma")]
-unsafe fn sq8_encode_row_avx2_unsafe_tail8(
-    row: &[f32],
-    inv_scale: &[f32],
-    c2: &[f32],
-    dst: &mut [u8],
-    i: &mut usize,
-) {
-    use std::arch::x86_64::*;
-    let zero = _mm256_setzero_ps();
-    let max255 = _mm256_set1_ps(255.0);
-    // SAFETY: caller-guaranteed `*i + 8 <= row.len()`.
-    unsafe {
-        let r = _mm256_loadu_ps(row.as_ptr().add(*i));
-        let inv = _mm256_loadu_ps(inv_scale.as_ptr().add(*i));
-        let c = _mm256_loadu_ps(c2.as_ptr().add(*i));
-        let q = _mm256_fmadd_ps(r, inv, c);
-        let q_clamped = _mm256_max_ps(_mm256_min_ps(q, max255), zero);
-        let q_i32 = _mm256_cvttps_epi32(q_clamped);
-        let lo = _mm256_castsi256_si128(q_i32);
-        let hi = _mm256_extracti128_si256::<1>(q_i32);
-        let packed_u16 = _mm_packus_epi32(lo, hi);
-        let packed_u8 = _mm_packus_epi16(packed_u16, packed_u16);
-        let dst_ptr = dst.as_mut_ptr().add(*i) as *mut i64;
-        std::ptr::write_unaligned(dst_ptr, _mm_cvtsi128_si64(packed_u8));
-        *i += 8;
-    }
-}
 
 /// Pass 2 of `build_subsection_streaming`: walk the input
 /// corpus chunk-by-chunk, assign each row to its centroid,
@@ -1744,221 +1301,6 @@ mod tests {
             rot_seed: 7,
             metric: Metric::L2Sq,
             rerank_codec: RerankCodec::Fp32,
-        }
-    }
-
-    /// Deterministic synthetic Sq8 encode inputs spanning a
-    /// realistic dynamic range: per-dim mins / maxes drawn from
-    /// distinct centroid offsets so `(x - offset) / scale` exercises
-    /// the full `[0, 255]` quantization range and the boundary
-    /// rounding behaviour where SIMD and scalar paths could
-    /// disagree by 1 ulp.
-    fn synth_sq8_inputs(dim: usize, seed: u64) -> (Vec<f32>, Vec<f32>, Vec<f32>) {
-        let mut state = seed.wrapping_mul(0x9E37_79B9_7F4A_7C15);
-        let mut next = || {
-            state = state
-                .wrapping_mul(6_364_136_223_846_793_005)
-                .wrapping_add(1_442_695_040_888_963_407);
-            ((state >> 33) as u32) as f32 / (u32::MAX as f32)
-        };
-        let mut scale = vec![0.0f32; dim];
-        let mut offset = vec![0.0f32; dim];
-        let mut row = vec![0.0f32; dim];
-        for d in 0..dim {
-            offset[d] = -2.0 + 4.0 * next();
-            let span = 0.1 + 4.0 * next();
-            scale[d] = span / 255.0;
-            // Mix of in-range, low-tail, and high-tail samples to
-            // exercise clamp + rounding boundaries.
-            let pick = next();
-            row[d] = if pick < 0.05 {
-                offset[d] - 0.5
-            } else if pick > 0.95 {
-                offset[d] + span + 0.5
-            } else {
-                offset[d] + span * next()
-            };
-        }
-        (row, scale, offset)
-    }
-
-    /// Scalar reference Sq8 encode using the exact same FMA + clamp
-    /// + truncating-cast sequence that `sq8_encode_row_*` use. This
-    /// is the byte-pattern oracle the SIMD parity tests assert
-    /// against — the on-disk Sq8 format is what this function
-    /// produces, regardless of which tier the runtime picks.
-    #[allow(clippy::manual_clamp)] // bit-match SIMD NaN handling
-    fn sq8_encode_row_reference(row: &[f32], inv_scale: &[f32], c2: &[f32], dst: &mut [u8]) {
-        debug_assert_eq!(row.len(), dst.len());
-        for d in 0..row.len() {
-            let q = row[d].mul_add(inv_scale[d], c2[d]);
-            let q_clamped = q.max(0.0).min(255.0);
-            dst[d] = q_clamped as u8;
-        }
-    }
-
-    /// Update-min/max parity: every tier converges on the same
-    /// per-dim (min, max) across a sweep of dims that hits the
-    /// 16-, 8-, and scalar-tail paths.
-    #[test]
-    fn update_min_max_simd_paths_match_scalar() {
-        let dims = [
-            1, 7, 8, 15, 16, 17, 24, 31, 32, 33, 47, 48, 96, 384, 512, 1023,
-        ];
-        for &dim in &dims {
-            let (row, _scale, _offset) = synth_sq8_inputs(dim, dim as u64 * 13);
-            let mut mn_ref = vec![f32::INFINITY; dim];
-            let mut mx_ref = vec![f32::NEG_INFINITY; dim];
-            for d in 0..dim {
-                if row[d] < mn_ref[d] {
-                    mn_ref[d] = row[d];
-                }
-                if row[d] > mx_ref[d] {
-                    mx_ref[d] = row[d];
-                }
-            }
-
-            for tier in ["wide", "avx2", "avx512"] {
-                let mut mn = vec![f32::INFINITY; dim];
-                let mut mx = vec![f32::NEG_INFINITY; dim];
-                match tier {
-                    "wide" => update_min_max_wide(&row, &mut mn, &mut mx),
-                    #[cfg(target_arch = "x86_64")]
-                    "avx2" if std::is_x86_feature_detected!("avx2") => {
-                        unsafe { update_min_max_avx2(&row, &mut mn, &mut mx) };
-                    }
-                    #[cfg(target_arch = "x86_64")]
-                    "avx512" if std::is_x86_feature_detected!("avx512f") => {
-                        unsafe { update_min_max_avx512(&row, &mut mn, &mut mx) };
-                    }
-                    _ => continue,
-                };
-                assert_eq!(mn, mn_ref, "tier {} min mismatch at dim {}", tier, dim);
-                assert_eq!(mx, mx_ref, "tier {} max mismatch at dim {}", tier, dim);
-            }
-        }
-    }
-
-    /// Sq8 SIMD encode parity: every tier produces byte-identical
-    /// codes to the FMA-based scalar reference across a sweep of
-    /// dims that hits all per-tier tail paths.
-    #[test]
-    fn sq8_encode_row_simd_paths_match_scalar() {
-        let dims = [
-            1, 7, 8, 15, 16, 17, 24, 31, 32, 33, 47, 48, 96, 384, 512, 1023,
-        ];
-        for &dim in &dims {
-            let (row, scale, offset) = synth_sq8_inputs(dim, dim as u64 * 17 + 1);
-            let consts = Sq8EncodeConsts::from_scale_offset(&scale, &offset);
-            let mut dst_ref = vec![0u8; dim];
-            sq8_encode_row_reference(&row, &consts.inv_scale, &consts.c2, &mut dst_ref);
-
-            // Wide tier — universally available.
-            let mut dst_wide = vec![0u8; dim];
-            sq8_encode_row_wide(&row, &consts.inv_scale, &consts.c2, &mut dst_wide);
-            assert_eq!(dst_wide, dst_ref, "wide path mismatch at dim {}", dim);
-
-            #[cfg(target_arch = "x86_64")]
-            if std::is_x86_feature_detected!("avx2") && std::is_x86_feature_detected!("fma") {
-                let mut dst_avx2 = vec![0u8; dim];
-                unsafe {
-                    sq8_encode_row_avx2(&row, &consts.inv_scale, &consts.c2, &mut dst_avx2);
-                }
-                assert_eq!(dst_avx2, dst_ref, "avx2 path mismatch at dim {}", dim);
-            }
-            #[cfg(target_arch = "x86_64")]
-            if std::is_x86_feature_detected!("avx512f") {
-                let mut dst_avx512 = vec![0u8; dim];
-                unsafe {
-                    sq8_encode_row_avx512(&row, &consts.inv_scale, &consts.c2, &mut dst_avx512);
-                }
-                assert_eq!(dst_avx512, dst_ref, "avx512 path mismatch at dim {}", dim);
-            }
-        }
-    }
-
-    /// Microbench: per-tier Sq8 encode throughput across the same
-    /// dim grid the AVX-512 / AVX2 distance microbenches use. Gated
-    /// `#[ignore]` so `cargo test --release` skips it; run via
-    /// `cargo test --release --lib sq8_encode_microbench --
-    /// --ignored --nocapture` to print a markdown table.
-    #[test]
-    #[ignore = "perf microbench, not a correctness gate"]
-    fn sq8_encode_microbench() {
-        use std::time::Instant;
-        let dims: &[usize] = &[128, 384, 768, 1024, 1536];
-        let iters: usize = 200_000;
-
-        println!("\n### Sq8 f32 → u8 encode per-tier ns / row (dim sweep)\n");
-        println!("| dim | scalar ns | wide ns | avx2 ns | avx512 ns |");
-        println!("|----:|----------:|--------:|--------:|----------:|");
-
-        for &dim in dims {
-            let (row, scale, offset) = synth_sq8_inputs(dim, dim as u64 * 23 + 5);
-            let consts = Sq8EncodeConsts::from_scale_offset(&scale, &offset);
-            let mut dst = vec![0u8; dim];
-
-            // Scalar reference — what the on-disk encode would
-            // be without any of this commit's SIMD work.
-            let t0 = Instant::now();
-            for _ in 0..iters {
-                sq8_encode_row_reference(&row, &consts.inv_scale, &consts.c2, &mut dst);
-                std::hint::black_box(&dst);
-            }
-            let scalar_ns = t0.elapsed().as_nanos() as f64 / iters as f64;
-
-            let t0 = Instant::now();
-            for _ in 0..iters {
-                sq8_encode_row_wide(&row, &consts.inv_scale, &consts.c2, &mut dst);
-                std::hint::black_box(&dst);
-            }
-            let wide_ns = t0.elapsed().as_nanos() as f64 / iters as f64;
-
-            #[cfg(target_arch = "x86_64")]
-            let avx2_ns =
-                if std::is_x86_feature_detected!("avx2") && std::is_x86_feature_detected!("fma") {
-                    let t0 = Instant::now();
-                    for _ in 0..iters {
-                        unsafe {
-                            sq8_encode_row_avx2(&row, &consts.inv_scale, &consts.c2, &mut dst);
-                        }
-                        std::hint::black_box(&dst);
-                    }
-                    Some(t0.elapsed().as_nanos() as f64 / iters as f64)
-                } else {
-                    None
-                };
-            #[cfg(not(target_arch = "x86_64"))]
-            let avx2_ns: Option<f64> = None;
-
-            #[cfg(target_arch = "x86_64")]
-            let avx512_ns = if std::is_x86_feature_detected!("avx512f") {
-                let t0 = Instant::now();
-                for _ in 0..iters {
-                    unsafe {
-                        sq8_encode_row_avx512(&row, &consts.inv_scale, &consts.c2, &mut dst);
-                    }
-                    std::hint::black_box(&dst);
-                }
-                Some(t0.elapsed().as_nanos() as f64 / iters as f64)
-            } else {
-                None
-            };
-            #[cfg(not(target_arch = "x86_64"))]
-            let avx512_ns: Option<f64> = None;
-
-            let fmt = |x: Option<f64>| match x {
-                Some(v) => format!("{:>7.1}", v),
-                None => "    n/a".to_string(),
-            };
-            println!(
-                "| {:>3} | {:>9.1} | {:>7.1} | {} | {} |",
-                dim,
-                scalar_ns,
-                wide_ns,
-                fmt(avx2_ns),
-                fmt(avx512_ns),
-            );
         }
     }
