+++ src/superfile/vector/mod.rs
@@ -18,3 +18,4 @@ pub mod reservoir;
 pub mod rotation;
 pub mod simd_dispatch;
 pub mod spill;
+pub mod sq8_simd;
